Seprate component for TodoItem
